# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/hr/attendance/{employeeId}/summary":  {
                                                                  "get":  {
                                                                              "operationId":  "AttendanceController_getSummary",
                                                                              "parameters":  [
                                                                                                 {
                                                                                                     "name":  "employeeId",
                                                                                                     "required":  true,
                                                                                                     "in":  "path",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 },
                                                                                                 {
                                                                                                     "name":  "month",
                                                                                                     "required":  true,
                                                                                                     "in":  "query",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 },
                                                                                                 {
                                                                                                     "name":  "year",
                                                                                                     "required":  true,
                                                                                                     "in":  "query",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 }
                                                                                             ],
                                                                              "responses":  {
                                                                                                "200":  {
                                                                                                            "description":  ""
                                                                                                        }
                                                                                            },
                                                                              "tags":  [
                                                                                           "Attendance"
                                                                                       ]
                                                                          }
                                                              },
...
                  "/api/ai/proposal-generator":  {
                                                     "post":  {
                                                                  "operationId":  "AiController_generateProposal",
                                                                  "parameters":  [

                                                                                 ],
                                                                  "requestBody":  {
                                                                                      "required":  true,
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/GenerateProposalDto"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                  "responses":  {
                                                                                    "201":  {
                                                                                                "description":  ""
                                                                                            }
                                                                                },
                                                                  "tags":  [
                                                                               "Ai"
                                                                           ]
                                                              }
                                                 },
...
                  "/api/ai/seo-audit":  {
                                            "post":  {
                                                         "operationId":  "AiController_auditWebsite",
                                                         "parameters":  [

                                                                        ],
                                                         "requestBody":  {
                                                                             "required":  true,
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/AuditWebsiteDto"
                                                                                                                                 }
                                                                                                                  }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "201":  {
                                                                                       "description":  ""
                                                                                   }
                                                                       },
                                                         "tags":  [
                                                                      "Ai"
                                                                  ]
                                                     }
                                        },
...
                  "/api/ai/email-writer":  {
                                               "post":  {
                                                            "operationId":  "AiController_writeEmail",
                                                            "parameters":  [

                                                                           ],
                                                            "requestBody":  {
                                                                                "required":  true,
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/WriteEmailDto"
                                                                                                                                    }
                                                                                                                     }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "201":  {
                                                                                          "description":  ""
                                                                                      }
                                                                          },
                                                            "tags":  [
                                                                         "Ai"
                                                                     ]
                                                        }
                                           },
                  "/api/ai/meeting-summary":  {
                                                  "post":  {
                                                               "operationId":  "AiController_summarizeMeeting",
                                                               "parameters":  [

                                                                              ],
                                                               "requestBody":  {
                                                                                   "required":  true,
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/SummarizeMeetingDto"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "201":  {
                                                                                             "description":  ""
                                                                                         }
                                                                             },
                                                               "tags":  [
                                                                            "Ai"
                                                                        ]
                                                           }
                                              },
                  "/api/ai/chat":  {
                                       "post":  {
                                                    "operationId":  "AiController_chat",
                                                    "parameters":  [

                                                                   ],
                                                    "requestBody":  {
                                                                        "required":  true,
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/AiChatDto"
                                                                                                                            }
                                                                                                             }
                                                                                    }
                                                                    },
                                                    "responses":  {
                                                                      "201":  {
                                                                                  "description":  ""
                                                                              }
                                                                  },
                                                    "tags":  [
                                                                 "Ai"
                                                             ]
                                                }
                                   },
...
              },
    "info":  {
                 "title":  "Trifusion-Dynamics Auth Service",
                 "description":  "The API description for Auth Service",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "GenerateProposalDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {

                                                                                  }
                                                               },
                                       "AuditWebsiteDto":  {
                                                               "type":  "object",
                                                               "properties":  {

                                                                              }
                                                           },
                                       "WriteEmailDto":  {
                                                             "type":  "object",
                                                             "properties":  {

                                                                            }
                                                         },
                                       "SummarizeMeetingDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {

                                                                                  }
                                                               },
                                       "AiChatDto":  {
                                                         "type":  "object",
                                                         "properties":  {

                                                                        }
                                                     },
...
                                   }
                   }
}